+++ src/Instructions.ts
@@ -8,7 +8,7 @@ interface InstructionsProps {
 /**
  * List of instructions available to the stack.
  * @author CasperSocio
- * @version 0.0.3
+ * @version 0.0.5
  * @since 0.0.1
  */
 export const Instructions: InstructionsProps = {
@@ -22,10 +22,14 @@ export const Instructions: InstructionsProps = {
   'PADDING_LEFT': new Instruction('PADDING_LEFT'),
   'PADDING_RIGHT': new Instruction('PADDING_RIGHT'),
   'RESET': new Instruction('RESET'),
+  'TEXT_ALIGN_CENTER': new Instruction('TEXT_ALIGN_CENTER'),
+  'TEXT_ALIGN_LEFT': new Instruction('TEXT_ALIGN_LEFT'),
+  'TEXT_ALIGN_RIGHT': new Instruction('TEXT_ALIGN_RIGHT'),
   'TEXT_DECORATION_ITALIC': new Instruction('TEXT_DECORATION_ITALIC'),
   'TEXT_DECORATION_STRONG': new Instruction('TEXT_DECORATION_STRONG'),
   'TEXT_DECORATION_UNDERLINE': new Instruction('TEXT_DECORATION_UNDERLINE'),
   'TEXT_TRANSFORM_CAPITALIZE': new Instruction('TEXT_TRANSFORM_CAPITALIZE'),
   'TEXT_TRANSFORM_LOWERCASE': new Instruction('TEXT_TRANSFORM_LOWERCASE'),
   'TEXT_TRANSFORM_UPPERCASE': new Instruction('TEXT_TRANSFORM_UPPERCASE'),
+  'WIDTH': new Instruction('WIDTH'),
 }