+++ src/classes/Element.ts
@@ -11,43 +11,58 @@ import {
   IInstruction,
   IStyleRules
 } from '../typings/interfaces'
-import { TContent, TInstructionName, TLogPrefix, TNumericRule } from '../typings/types'
+import {
+  TContent,
+  TInstructionName,
+  TLogPrefix,
+  TNumericRule,
+  TTag
+} from '../typings/types'
 import { Instruction } from './Instruction'
 import { InstructionStack } from './InstructionStack'
 
 /**
+ * The valid props to pass a new Element instance.
+ */
+type InputProps = TTag | IElementList | IElementParagraph
+
+/**
  * The Element class.
  * @author CasperSocio
- * @version 0.0.3
+ * @version 0.0.5
  * @since 0.0.1
  */
  export class Element {
+  private _activeStyles: IInstruction[]
   private _content: TContent
   private _stack
-  private _log: string[]
+  private _log: [TLogPrefix, string][]
   private _style: IStyleRules
   private _tag
 
-  constructor({
-    content,
-    tag
-  }: IElementList | IElementParagraph) {
-    this._stack = new InstructionStack()
+  constructor(input: InputProps) {
+    this._activeStyles = []
     this._log = []
+    this._stack = new InstructionStack()
     this._style = {}
-    this._tag = tag
 
-    // Initialize this._content
-    switch (this._tag) {
-      case 'ol':
-      case 'ul':
-        this._content = content || []
-        break
-      case 'p':
-      default:
-        this._content = content || ''
-        break
+    if (typeof input === 'string') {
+      this._tag = input
+      this._content = ''
+    } else {
+      this._tag = input.tag
+      switch (this._tag) {
+        case 'ol':
+        case 'ul':
+          this._content = input.content || []
+          break
+        case 'p':
+        default:
+          this._content = input.content || ''
+          break
+      }
     }
+
   }
 
   public set content(content: TContent) {
@@ -58,6 +73,27 @@ import { InstructionStack } from './InstructionStack'
   public set style(rules: IStyleRules) {
     this.log('LOG', 'Setting new style rules')
     this._style = rules
+  }
+
+  /**
+   * Adds a new instruction to this._activeStyles.
+   * @author CasperSocio
+   * @version 0.0.5
+   * @param instruction The instruction to add
+   * @since 0.0.5
+   * @private
+   */
+  private activateStyle(instruction: IInstruction) {
+    let exists: boolean = false
+    this._activeStyles.forEach(style => {
+      if (style.name === instruction.name) {
+        exists = true
+      }
+    })
+    if (!exists) {
+      this.log('LOG', 'Activating ' + instruction.name)
+      this._activeStyles.push(instruction)
+    }
   }
 
   /**
@@ -113,7 +149,7 @@ import { InstructionStack } from './InstructionStack'
    * @since 0.0.3
    * @private
    */
-  private addToStack(instruction: IInstruction) {
+  private addToStack(instruction: IInstruction, value?: number | string) {
     this.log('INS', [
       StyleTextDecoration.strong,
       StyleColor.green,
@@ -123,13 +159,13 @@ import { InstructionStack } from './InstructionStack'
       ' ' + instruction.name + ' ',
       StyleUtilities.reset
     ].join(''))
-    this._stack.add(instruction)
+    this._stack.add(instruction, value)
   }
 
   /**
    * Adds new styling instructions to the stack.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @since 0.0.1
    * @private
    */
@@ -179,64 +215,89 @@ import { InstructionStack } from './InstructionStack'
     // Padding-right
     this.addNumberValueInstruction(this._style.paddingRight, Instructions.PADDING_RIGHT)
 
+    // Text-align
+    this._style.textAlign && this.addToStack(Instructions[`TEXT_ALIGN_${this._style.textAlign.toUpperCase()}`])
+
     // Text-decoration
-    if (this._style.textDecoration) {
-      switch (this._style.textDecoration) {
-        case 'italic':
-          this.addToStack(Instructions.TEXT_DECORATION_ITALIC)
-          break
-        case 'strong':
-          this.addToStack(Instructions.TEXT_DECORATION_STRONG)
-          break
-        case 'underline':
-          this.addToStack(Instructions.TEXT_DECORATION_UNDERLINE)
-          break
-        default:
-          break
+    this._style.textDecoration && this.addToStack(Instructions[`TEXT_DECORATION_${this._style.textDecoration.toUpperCase()}`])
+
+    // Text-transform
+    this._style.textTransform && this.addToStack(Instructions[`TEXT_TRANSFORM_${this._style.textTransform.toUpperCase()}`])
+
+    // Width
+    if (this._style.width !== undefined) {
+
+      // Find the total length of content + padding
+      let preWidth: number = this._content.length
+      let hasBackgroundPaddingLeft: boolean,
+          hasBackgroundPaddingRight: boolean
+
+      hasBackgroundPaddingLeft = (
+        this._style.backgroundColor && !this._style.paddingLeft
+        ? true
+        : false
+      )
+      hasBackgroundPaddingRight = (
+        this._style.backgroundColor && !this._style.paddingRight
+        ? true
+        : false
+      )
+
+      if (hasBackgroundPaddingLeft || hasBackgroundPaddingRight)
+        preWidth++
+      if (this._style.paddingLeft && typeof this._style.paddingLeft === 'number')
+        preWidth += this._style.paddingLeft
+      if (this._style.paddingRight && typeof this._style.paddingRight === 'number')
+        preWidth += this._style.paddingRight
+
+      // Only add width if (width > content.length + padding)
+      if (preWidth < this._style.width) {
+        // Add 'WIDTH' instructions
+        this.addToStack(Instructions.WIDTH, this._style.width - preWidth)
       }
     }
 
-    // Text-transform
-    if (this._style.textTransform) {
-      switch (this._style.textTransform) {
-        case 'capitalize':
-          this.addToStack(Instructions.TEXT_TRANSFORM_CAPITALIZE)
-          break
-        case 'lowercase':
-          this.addToStack(Instructions.TEXT_TRANSFORM_LOWERCASE)
-          break
-        case 'uppercase':
-          this.addToStack(Instructions.TEXT_TRANSFORM_UPPERCASE)
-          break
-        default:
-          break
-      }
-    }
     this.log('ACT', 'APPLYING STYLES [END]')
+  }
+
+  /**
+   * Removes an instruction from this._activeStyles.
+   * @author CasperSocio
+   * @version 0.0.5
+   * @param instruction The instruction to remove
+   * @since 0.0.5
+   * @private
+   */
+  private deactivateStyle(instruction: IInstruction) {
+    let newActiveStyles: IInstruction[] = []
+    this._activeStyles.forEach(style => {
+      if (style === instruction) {
+        this.log('LOG', 'Deactivating ' + instruction.name)
+      } else {
+        newActiveStyles.push(style)
+      }
+    })
+    this._activeStyles = newActiveStyles
   }
 
   /**
    * Adds a new log entry.
    * @author CasperSocio
-   * @version 0.0.1
+   * @version 0.0.5
    * @param prefix What type of log to use
    * @param msg The string message
    * @since 0.0.1
    * @private
    */
   private log(prefix: TLogPrefix, msg: string) {
-    if (prefix === 'ACT') {
-      this._log.push(`${StyleColor.blue}[${prefix}] ${msg}${StyleUtilities.reset}`)
-    } else {
-      this._log.push(`[${prefix}] ${msg}`)
-    }
+    this._log.push([prefix, msg])
   }
 
   /**
    * Parses the instruction stack and
    * generates the final output string.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @since 0.0.1
    * @private
    */
@@ -249,32 +310,70 @@ import { InstructionStack } from './InstructionStack'
       this.log('PAR', `${instruction.name}`)
 
       switch (instruction.name) {
+        case 'BACKGROUND_COLOR':
+        case 'COLOR':
+        case 'TEXT_DECORATION_ITALIC':
+        case 'TEXT_DECORATION_STRONG':
+        case 'TEXT_DECORATION_UNDERLINE':
+          if (instruction.value) {
+            output.push(String(instruction.value))
+            this.activateStyle(instruction)
+          }
+          break
+        
         case 'CONTENT':
           if (instruction.value) {
             if (this._tag === 'ol' || this._tag === 'ul') {
               output.push(instruction.value + '\n')
-              break
+            } else {
+              output.push(String(instruction.value))
             }
-            output.push(instruction.value)
           }
+          this.deactivateStyle(Instructions.COLOR)
+          this.deactivateStyle(Instructions.TEXT_DECORATION_ITALIC)
+          this.deactivateStyle(Instructions.TEXT_DECORATION_STRONG)
+          this.deactivateStyle(Instructions.TEXT_DECORATION_UNDERLINE)
+          output.push(this.updateStyles())
           break
 
         case 'TEXT_TRANSFORM_CAPITALIZE':
-          this._stack.findAndReplaceValue('CONTENT', value => (
-            value.split(' ').map(word => word[0].toUpperCase() + word.slice(1)).join(' ')
-          ))
+          this._stack.findAndReplaceValue({
+            name: 'CONTENT',
+            callback: (value: string) => (
+              value.split(' ').map(word => word[0].toUpperCase() + word.slice(1)).join(' ')
+            )
+          })
           break
 
         case 'TEXT_TRANSFORM_LOWERCASE':
-          this._stack.findAndReplaceValue('CONTENT', value => value.toLowerCase())
+          this._stack.findAndReplaceValue({
+            name: 'CONTENT',
+            callback: (value: string) => value.toLowerCase()
+          })
           break
 
         case 'TEXT_TRANSFORM_UPPERCASE':
-          this._stack.findAndReplaceValue('CONTENT', value => value.toUpperCase())
+          this._stack.findAndReplaceValue({
+            name: 'CONTENT',
+            callback: value => value.toUpperCase()
+          })
+          break
+
+        case 'RESET':
+          this.deactivateStyle(Instructions.BACKGROUND_COLOR)
+          output.push(this.updateStyles())
+          break
+
+        case 'WIDTH':
+          if (instruction.value) {
+            for (let i = 1; i < instruction.value; i++) {
+              output.push(' ')
+            }
+          }
           break
       
         default:
-          instruction.value && output.push(instruction.value)
+          instruction.value && typeof instruction.value === 'string' && output.push(instruction.value)
           break
       }
     })
@@ -323,24 +422,68 @@ import { InstructionStack } from './InstructionStack'
   /**
    * Prints instruction stack.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @since 0.0.2
    * @public
    */
   public showInstructions() {
-    console.log(this._stack.stack)
+    console.log('[')
+    console.group()
+    this._stack.stack.forEach(Instruction => {
+      console.log(`{ ${StyleColor.green + Instruction.name + StyleUtilities.reset} }`)
+    })
+    console.groupEnd()
+    console.log(']')
   }
 
   /**
    * Prints all log entries.
    * @author CasperSocio
-   * @version 0.0.1
+   * @version 0.0.5
    * @since 0.0.1
    * @public
    */
   public showLog() {
+    let tabSize: number = 0
     this._log.forEach(log => {
-      console.log(log)
+      switch (log[0]) {
+        case 'ACT':
+          if (log[1].includes('[END]') && tabSize > 0) {
+            console.groupEnd()
+            tabSize--
+          }
+          console.log(`${StyleColor.blue}[${log[0]}] ${log[1] + StyleUtilities.reset}`)
+          if (log[1].includes('[START]')) {
+            console.group()
+            tabSize++
+          }
+          break
+
+        case 'PAR':
+          console.log(`${StyleColor.yellow}[${log[0]}] ${log[1] + StyleUtilities.reset}`)
+          break
+      
+        default:
+          console.log(`[${log[0]}] ${log[1]}`)
+          break
+      }
     })
+  }
+
+  /**
+   * Resets formatting and re-applies styling.
+   * @author CasperSocio
+   * @version 0.0.5
+   * @returns The updated formatting as a string
+   * @since 0.0.5
+   * @private
+   */
+  private updateStyles() {
+    this.log('LOG', 'Updating styles')
+    let updatedStyles: string[] = [StyleUtilities.reset]
+    this._activeStyles.forEach(style => {
+      style.value && updatedStyles.push(String(style.value))
+    })
+    return updatedStyles.join('')
   }
 }