+++ src/classes/InstructionStack.ts
@@ -15,23 +15,43 @@ export class InstructionStack {
   /**
    * Adds a new instruction to the stack.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @param instruction The new instruction to add
    * @since 0.0.3
    * @public
    */
-  add(instruction: IInstruction) {
+  add(instruction: IInstruction, value?: number | string) {
+    if (value) {
+      instruction.value = value
+    }
     this._stack.push(instruction)
   }
 
-  findByName(name: TInstructionName) {
-    let res: IInstruction[] = []
-    this._stack.forEach(instruction => {
-      if (instruction.name === name) {
-        res.push(instruction)
+  /**
+   * Checks if the stack contains a given instruction.
+   * @author CasperSocio
+   * @version 0.0.5
+   * @param name The instruction.name to look for
+   * @returns boolean
+   * @since 0.0.5
+   * @public
+   */
+  contains(name: TInstructionName) {
+    for (let i in this._stack) {
+      if (this._stack[i].name === name) {
+        return true
       }
-    })
-    return res
+    }
+    return false
+  }
+
+  findByName(name: TInstructionName) {
+    for (let i in this._stack) {
+      if (this._stack[i].name === name) {
+        return this._stack[i]
+      }
+    }
+    return false
   }
 
   /**
@@ -55,19 +75,49 @@ export class InstructionStack {
   /**
    * Finds matching instructions by name and assigns a new value.
    * @author CasperSocio
-   * @version 0.0.3
+   * @version 0.0.5
    * @param name The instruction target name
    * @param value The new value to assign instruction
+   * @param callback The callback function
    * @since 0.0.3
    * @public
    */
-  findAndReplaceValue(name: TInstructionName, callback: (value: string) => string) {
+  findAndReplaceValue({ name, value, callback }: {
+    name: TInstructionName
+    value?: number | string
+    callback?: (newValue: string) => string
+  }): void {
     this._stack = this._stack.map(instruction => {
       if (instruction.name === name) {
-        instruction.value = callback(instruction.value || '')
+        if (value !== undefined) {
+          instruction.value = value
+        }
+        else if (callback !== undefined) {
+          if (typeof instruction.value === 'string') {
+            instruction.value = callback(instruction.value || '')
+          }
+        }
       }
       return instruction
     })
+  }
+
+  /**
+   * Finds the stack index of an instruction.
+   * @author CasperSocio
+   * @version 0.0.5
+   * @param name Name of the instruction
+   * @returns The index number
+   * @since 0.0.5
+   * @public
+   */
+  indexOf(name: TInstructionName) {
+    for (let i = 0; i < this._stack.length; i++) {
+      if (this._stack[i].name === name) {
+        return i
+      }
+    }
+    return -1
   }
 
   /**