+++ src/typings/types.ts
@@ -4,7 +4,7 @@ import {
 } from './interfaces'
 
 /** The available color options */
-export type TColor = 'black' | 'blue' | 'green' | 'red' | 'white' | 'yellow'
+export type TColor = 'black' | 'blue' | 'cyan' | 'green' | 'magenta' | 'red' | 'white' | 'yellow'
 
 export type TContent = string | string[]
 
@@ -21,12 +21,16 @@ export type TInstructionName =
   | 'PADDING_LEFT'
   | 'PADDING_RIGHT'
   | 'RESET'
+  | 'TEXT_ALIGN_CENTER'
+  | 'TEXT_ALIGN_LEFT'
+  | 'TEXT_ALIGN_RIGHT'
   | 'TEXT_DECORATION_ITALIC'
   | 'TEXT_DECORATION_STRONG'
   | 'TEXT_DECORATION_UNDERLINE'
   | 'TEXT_TRANSFORM_CAPITALIZE'
   | 'TEXT_TRANSFORM_LOWERCASE'
   | 'TEXT_TRANSFORM_UPPERCASE'
+  | 'WIDTH'
 
 export type TLogPrefix = 'ACT' | 'INS' | 'LOG' | 'PAR'